test(codegen): split generator tests into focused modules
Replace the 620-line ``tests/scripts/test_generate_tool_versions.py``
with a directory mirroring the generator package: ``test_seed.py``,
``test_inputs.py``, ``test_outputs.py``, ``test_main.py``,
``test_real_repo.py``. Move ``gen``, ``fake_repo``, and ``retargeted_gen``
fixtures into ``conftest.py``. The new ``retargeted_gen`` fixture
replaces the in-test ``_retarget`` helper so end-to-end tests just
declare the fixture.

Update ``test_scripts_readme_coverage`` to skip files inside private
packages (any path component starting with ``_``); the generator's
``_generator/*.py`` modules are implementation detail of the documented
``generate-tool-versions.py`` entry script.
